Covnert to double outside of the get_ncdata_from_series function
I use the `get_ncdata_from_series` in other libraries/programs that
do need support for time being an integer, so I've left it up to whatever
calls that to convert the result to the correct dtype itself.

+++ pocean/dsg/timeseries/om.py
@@ -137,7 +137,7 @@ class OrthogonalMultidimensionalTimeseries(CFDataset):
 
             attributes = dict_update(nc.nc_attributes(axes, daxes), kwargs.pop('attributes', {}))
 
-            time[:] = get_ncdata_from_series(t, time)
+            time[:] = get_ncdata_from_series(t, time).astype('f8')
 
             # Create vars based on full dataframe (to get all variables)
             for c in data_columns:
